feat: 完成用户模块所有API，准备合并main分支最新代码
